Make Yarn based Hadoop docker distribution
The current solution creates a Yarn based hadoop docker distribution. It
consists the basic configuration, not additional cofiguration
configuration is possible at the time. A user can start a docker
container and log into it to interract with hadoop through the `hdfs dfs`
command. All the Web UI forms are available if the user forwards needed
ports from the container to their own computer. A docker-compose
demonstrates how to use the image.

Signed-off-by: Nikita Eshkeev <neshkeev@yandex.ru>